third_party/blink/web_tests/external/wpt/url/url-setters-a-area.window_include=javascript-expected.txt

This is a testharness.js-based test.
[PASS] Loading data…
[PASS] <a>: Setting <javascript:alert(1)>.protocol = 'defuse'
[PASS] <area>: Setting <javascript:alert(1)>.protocol = 'defuse'
[PASS] <a>: Setting <javascript:alert(1)>.username = 'wario'
[PASS] <area>: Setting <javascript:alert(1)>.username = 'wario'
[FAIL] <a>: Setting <javascript://x/>.username = 'wario'
  assert_equals: expected "javascript://wario@x/" but got "javascript://x/"
[FAIL] <area>: Setting <javascript://x/>.username = 'wario'
  assert_equals: expected "javascript://wario@x/" but got "javascript://x/"
[FAIL] <a>: Setting <javascript://x/>.password = 'bowser'
  assert_equals: expected "javascript://:bowser@x/" but got "javascript://x/"
[FAIL] <area>: Setting <javascript://x/>.password = 'bowser'
  assert_equals: expected "javascript://:bowser@x/" but got "javascript://x/"
[FAIL] <a>: Setting <javascript://x/>.port = '12'
  assert_equals: expected "javascript://x:12/" but got "javascript://x/"
[FAIL] <area>: Setting <javascript://x/>.port = '12'
  assert_equals: expected "javascript://x:12/" but got "javascript://x/"
[PASS] <a>: Setting <javascript:alert(1)>.hash = 'castle'
[PASS] <area>: Setting <javascript:alert(1)>.hash = 'castle'
Harness: the test ran to completion.
